Fix media_grants for shared package names
See bug for context. If packages use a shared package name, `getNameForUid`
may not return the correct package name for the purposes of query file
access. Instead, we use one of the valid packages in `getPackagesForUid` since
the underlying query logic checks for file access for any of the
packages returned in this list (in the event of shared packages).

Bug: b/289987807
Test: atest MediaGrantsTest
Change-Id: I45cf438a4b687ca7e5e1642a6cf8cb5cf59c220e