Fix uncaught IllegalArgumentException onPackageOrphaned
This is thrown when creating LocalCallingIdentity with an Orphaned packages UID
(in case of additional shared packages that need to have media grants
cleaned up) and ultimately causes the ongoing database transaction to get
rolled back if there isn't any remaining packages with the given UID on the
system.

In the event no additional shared packages are found, just use the
package name of the uninstalled package to proceed.

Bug: b/301411633
Test: atest MediaGrantsTest
Change-Id: I248e9d45bf754e8e0d98bfcb40d9b5b3a77a732c