Reduce test flake rate by making tests more deterministic
Currently when test config updates, the test config listeners are
executed on different threads. This leads to a non-deterministic
behaviour where the sequence of statements running in a test can
change. This has resulted in test flakes.

By running the test config listeners in a synchronised manner, we can
reduce the non-deterministic behaviour in tests and write less flaky
tests.

Bug: 299628853
Test: atest PickerSyncControllerTest
Change-Id: I699169055de552f58425438d6d9b67f5339966ca