Update backup setup code
attachVolume() gets called for each volume multiple times. It gets
called in onCreate, setting up backup during this stage would lead to
long wait in getting FuseDaemon thread for external_primary volume. Wait
of more than 10s then leads to an ANR. Modified code to accept volume
mount state in attachVolume call. When volume is mounted and called from
ExternalStorageService, we pass the state and only in this case we setup
the backup code. In this case fuse daemon thread should already be
created in onStartSession() method of ExternalStorageService and there
should not be a long wait to acquire FuseDaemon thread.

Bug: 304230629
Test: Verified manually from logs
Change-Id: I34ff4248abebe5ec037d4df37cfb1250c2f9ccad