Add null checks in TabFragment classes
Method Fragment#getActivity() can return null if the Fragment is not
associated with an activity. This is specially a problem in registered
listeners because it's possible that by the time they are executed,
there state of the fragment is changed and it is no longer attached to
an activity. In such rare cases, the exception should be caught, logged
and swallowed so that PhotoPicker process does not crash.

Bug: 305003800
Test: atest ActiveProfileButtonTest
Test: atest AlbumsTabTest
Test: atest PhotosTabTest
Change-Id: Ib5e61db2a8d1aee24b6839439d773ea7921c1b7f