Introduce Ordered Selection in PhotoPicker
Create a new TextView to display the order, which is maintained in the
Selection data in the picker frontend. When the user finishes, the
results are sorted by the selection order, and returned to the calling
app.

A new boolean extra EXTRA_PICK_ORDERED_IMAGES is introduced to gate this
feature. The feature is not allowed in GET_CONTENT or in the Permission
Mode.

Demo: https://drive.google.com/file/d/1E5HQJzD9czZD2gWq0AEhABG7Q-pqeFzY/view?usp=drive_link

Test: atest CtsPhotoPicker & atest com.android.providers.media.photopicker & manual tests (see demo)
Bug: 287661737
Change-Id: I639f958027cf4f0b758f232ae1e4ffe0d32db910