Init banner manager only once in a picker session
When changing the activity config like orientation to landscape /
portrait, the activity recreates. This currently causes the banner
manager to init again and hides the already showing banner.

With this change, we'll only init the banner manager once in a picker
session during the first on-create, post which, we only reset the banner
manager but not init.

Test: Manual
Bug: 303883680
Change-Id: Ia28fd09013fefc6ff26271d61deda67bf27dde68