Enable FUSE_CAP_PARALLEL_DIROPS
Allow lookup() and readdir() requests to issued concurrently for the
same directory without blocking.

Without this flag, when users issues lookup() and readdir() at the same
time, the kernel send the requests to FUseDaemon sequentially. This can
be problematic when the daemon's lookup handler gets stuck for some
reason. In such a case, the following readdir() will never be processed
and the kernel will panic if hung_task_panic is set.

Bug: 312226503
Test: Build locally
Change-Id: I32591b1cbe2878b0f959b1a1b2c11de8f5022dc4