Dispatch a ui refresh notification when a full sync is started.
This fixes the edge case of certain UI elements (Banners) not being
refreshed in the event of a collection-id change from the cloud
provider.

Now, when a full sync is known to be needed, PickerSyncController will
send a notification to any active running Picker activities to refresh
their UIs as data may be stale / incorrect.

Bug: b/313641265
Test: Manual w/ Pixel 7
Change-Id: I69feff34f330a0daece9454ea7cabd89e180facb