[Private Space] Refactor PhotoPickerActivity to setup dynamic number of Profiles
Refactored and Introduced new apis in order to make current photopicker
compatable with dynamic number of profiles. Pre-exiting apis were
designed to handle the fix number of two profiles (personal and managed)
but with new set of apis, the system is no more limited to personal and
managed profiles. The current system is designed to handle not only the
private space but any new profile that may be introduced in the future.

Bug: b/304436925

Test: `atest
com.android.providers.media.photopicker.viewmodel.PickerViewModelTest`

Change-Id: I0310dc8b9d2a3367c3c62f308e28af5ae9b44658