Update AsyncContentProvider timeout to 3 minutes.
The 5 second timeout is too low and is causing invalid FileNotFound
errors for devices with slower network or when downloading bigger files.

The Photopicker preloader will now block all file selection on
successful download.

3 minutes was chosen as a new timeout to stay under the system threshold
of terminating stuck threads (thought to be 10m), but to allow more time
for downloads from the CMP to complete.

Test: Manually, with Apps, Chrome browser for web apps
Bug: b/318460025
Change-Id: I269f9924b46adc51d264886e329cee0715b7fde4