Add EndlessWorker to PickerSyncManager.
EndlessWorker is a no-op worker that continually enqueues itself to
prevent the androidx.work JobScheduler from being disabled.

PackageManager sends a PACKAGE_CHANGED broadcast whenever components are
enabled and disabled. Since MediaProvider enqueues a worker as the result
of a file insert, this causes a PACKAGE_CHANGED broadcast when
WorkManager enables itself. By enqueueing a permanent worker it prevents
WorkManager from becoming disabled in the first place, and circumvents
this PACKAGE_CHANGED broadcast.

This was the recommended mitigation after speaking with WorkManager
team, and hopefully is only temporarily neccessary as they're aware of
the issue, but we want to mitigate the system performance concerns with
file inserts waking up a bunch of system services that listen to
PACKAGE_CHANGED.

The WorkManager component is androidx.work.impl.background.systemalarm.RescheduleReceiver.

Bug: b/319082111
Test: atest PickerSyncManagerTest
Change-Id: I6294e271cf1eab64f3cb28c0503439771791e899