Fix the ANR in PhotoPicker when any profile turned off or removed in the background
Fixed the ANR in photopicker when any profile turned off or removed in
the background. Now if user remove or turn off the profile that is
currently selected in PhotoPicker running in the background, the
photopicker will switch to the context user successfully withour any
crashing.

Bug: b/308901635
Test: Open PhotoPicker in Personal user and select any other profile
apart from personal. Now keep the photopicker running in the background
and turn off or remove the selected profile. When you return to picker
again, the picker should get switch to context user without any crash.

Change-Id: Ib524281a8a385c244652affbff50c4d629358c3f