Removing guava library to fix memory bloat.
tl;dr
A regression caused memory increases across Android processes. This was traced to an unexpectedly large increase in the framework-pdf.jar file.  The issue stemmed from a dependency on the Guava library, causing the inclusion of a large 'publicsuffix' file. I resolved this by creating a new utility class to replace Guava functions, eliminating the dependency and dramatically reducing the framework-pdf.jar file size.
http://shortn/_cnffWuCKh4

Test: m
Bug: 325462880
Change-Id: I1c4c316d67d44f41582765d509faf965c87bf723