Skip scanning MyFiles/Download in ARC
In ARC, /storage/0000000000000000000000000000CAFEF00D2019 is a mount
point of a StubVolume for ChromeOS "My files" (directory for
locally-saved user files), and
/storage/0000000000000000000000000000CAFEF00D2019/Downloads is the
ChromeOS Downloads directory. The contents of this directory is
identical to that of the Android's Download directory
(/storage/emulated/<user-id>/Download). To avoid scanning the same file
in these directories twice, this patch adds logic to skip scanning the
files in /storage/0000000000000000000000000000CAFEF00D2019/Downloads.

Original divergence: ag/21171318

Bug: 304369069
Test: On betty-arc-v, test that media files in MyFiles/Downloads/,
  MyFiles/Downloads/foo/ and MyFiles/Downloads-foo/ are scanned only
  once each.

Change-Id: I945cec610c3f2d33141331d1ea44e7b915aea9f2