Change for generating libpdfclient.so for 32 bit
All of pdf CTS tests are failing for a specific abi 'armeabi-v7a'
because 'libpdfclient.so' is not getting loaded for this specific ABI.
The reason for this is that the libpdfclient.so is not getting created for 32 bit. We need this change so that libpdfclient.so gets generated for both 32 bit and 64 bit.
See this doc for more details: https://docs.google.com/document/d/1_luwm1ktgfhibsnlA_557vEmNS95MLnE-xBQCjUTjec/edit?usp=sharing

Bug: 325628733
Test: Build and run cts tests
Change-Id: Ifa5ed88d78d4c851f3ed93725ba4e8459e51bb8b