Prevent insertion in other users storage volumes
Don't allow file insertion in other users storage volumes. This was
already handled if DATA was explicitly set in content values, but was
allowed if DATA was generated based on other values like RELATIVE_PATH
and DISPLAY_NAME. Insertion of files in other users storage volumes can be used by malicious
apps to get access to other users files, since the same file would exist in both users
MP databases which would lead to MP falsely assuming that the user has access to this file
while that's not true.

Bug: 294406604
Test: atest MediaProviderTests
Change-Id: I0a469d3bc4880903e749e66073bc4a36a8d7e12d