Fix for text getting cut off in Russian.
Changed the layout from FrameLayout to LinearLayout with the two buttons
having layout_weight in the ration 2:1. Text in both the buttons are set
to be of maxLines 1 and ellipsized to end for the `View selected` button
and to the middel for the `Add (n)` button.

Bug: 282883501
Bug: 313613936
Test: Manual
Screenshot: http://screen/4erUEXVJX4BwwqP
Change-Id: I8ccf4e0ec7f4acba1a8b6dcf9cc37b11f65b7664
Merged-In: I8ccf4e0ec7f4acba1a8b6dcf9cc37b11f65b7664