Block taking recents screenshot if private profile available
Photopicker shows separate tab when private profile is on.
User can move photopicker to the background and then lock private
profile. Recents preview screen shows screenshot
of the photopicker at the last moment it was in foreground
(when the space was unlocked). As a result Recents preview
can leak whether private profile exists on device.
This change sets setting to show blank
screen in the recents when private profile is avaialble.

Bug: 329241748
Test: manual, https://buganizer.corp.google.com/issues/329241748#comment3
Change-Id: Ic0848815565df4ad017d60972d4401e2135d491d