Fix crash in the DataLoader thread
In PickerViewModel, currently the list of media items can contain
an unmodifiable list. When new data is fetched asynchronously
from the DataLoader thread and the new items are added to the
media list, if the media list is an unmodifiable list, it throws
an UnsupportedOperationException.

This change makes sure that we always have a modifiable/mutable media
list held in PickerViewModel.

This exception also contributed to the flakiness of
CloudPhotoPickerTest#testProviderSwitchSuccess

Bug: 332843606
Test: atest CloudPhotoPickerTest
Change-Id: Ib40b987e016f05993ec2ddd4090784dab7c816d7