Added mts-mediaprovider tests suite to pdfclient_test
Followed the steps mentioned in the doc (http://go/mainline-test-suite#how-to-add-my-tests-to-mts) to include pdfclient in MTS. We needed to apply MainlineModuleTestModuleControlloer to our tests by adding the lines mentioned here (go/mainline-test-suite#for-optional-mainline-modules-in-android-q-and-all-other-mainline-modules-introduced-in-android-r) in the test configuration, so created a test config file for pdfclient_test and added the required details.
Verified the tests are passing successfully.

Bug: 333987974
Test: atest pdfClient_test
Change-Id: Ia7bf11ba4d82b5e3f7709f19f62d0aab31797569