Return a null pageSelction object from jni if the List of bounds is null for a page selection.
Note: The root cause of this problem is that JNI returns an empty list of bounds if we un-select a particular text till the selection reaches an empty area. Ideally for a valid PageSelection object list of bounds (list of rects coordinates for the selected area) can not be a empty list or null. So updating the JNI layer such that when the list of bounds is empty list (no text is selected or text is unselected to an empty area) we return a null PageSelection object instead.

Test: atest CtsPdfModuleTestCases
BUG: 334805330
Change-Id: I39623279624c96989480303b5a5ef3b30dfb91e0