FuseDaemon: return EXDEV when fuse files are renamed into fuse-bpf root
It is allowed to rename from the fuse path to the root, data or obb,
of the fuse-bpf path, however, it finally returns EBADF from the kernel
since the file belongs to a fuse-bpf path not having any backing
information. For the sake of compatibility, we just return EXDEV, the same
as the case with renaming a file to a sub-directories within the fuse-bpf
path (i.e., copy & unlink).

Bug: 336656707
Test: N.A.
Change-Id: I38a4968de0997f53a06d83594e85eb8fe94d2681
Merged-In: I38a4968de0997f53a06d83594e85eb8fe94d2681