Modify selection count calculation based on picker mode.
When picker is used in ACTION_USER_SELECT_IMAGES_FOR_APP mode the total count of
selection shall be more than the max limit set since the selections are
merged togther from previous session.

But for the case of PICK_IMAGES the total items selected i.e. current
selection + preGranted items should be the max limit passed in the
intent.

Test: atest SelectionTest.java

Bug: 339151767

Change-Id: I5a0d2d3383fc72313dec3144ce5f038711a65019