Pass non-null values to CMP API params marked as @NonNull
In Java CMPs, if null is passed to a @NonNull API param, a warning is
generated but there are no exceptions.

In contrast, in Kotlin CMPs, if null is passed to a @NonNull value, a
runtime exception occurs.

This change fixes this problem for extras received in
onGetMediaCollectionInfo method and onCreateCloudMediaSurfaceController
method in CloudMediaProvider.

Bug: 343358319
Test: -
Change-Id: I5cfb2c1f4fb3aed341b41ec40843c2794ee83a68