Add flag to control enabling the photopicker DeviceConfigReceiver
Wear devices don't use photopicker, but we are seeing the photopicker launching during boot. It launches because it is currently listening for an intent to tell it to check DeviceConfig to see if "enable_modern_picker" is set. Since Wear doesn't use photopicker, we would like to not have this work done on boot, and this change adds a flag to disable the DeviceConfigReceiver so that it will not be launched at boot.

The reason I opted to make this change specifically is because we're actually not 100% sure if photopicker is never used by any app. It appears that, if this photopicker app does not get enabled via this DeviceConfig flag, then the "non-modern" version will take over instead, so we do not lose the ability to launch the photopicker due to this change. Furthermore, when the legacy photopicker becomes EOL, then there should be a smooth switchover because the modern photopicker should no longer require a boot-time query to DeviceConfig to determine whether it's enabled (because it would always be enabled).

bug: 367926773
test: manually verify changing flag has intended effect on com.google.android.photopicker at boot
Flag: NONE Additional Resource Overlay Path
Change-Id: I60585aac4db1b6e84a83490291332db1e5dfb9c1