Implement the open transcoded file flow of Photopicker Transcoding
The open file flow will be redirected to PhotoPickerTranscodeHelper if
the URI is a photo picker transcoded URI. The helper class will first
try to open the pre-transcoded file. If the file is not existed, which
might be due to cache cleaning up for space, the helper class will try
to open the source file as fallback strategy.

For details see: go/photopicker-hdr-video-transcoding.

Test: manual test & PickerUriResolverTest
Bug: 365461942
Flag: com.android.providers.media.flags.enable_photopicker_transcoding
Change-Id: I1cc4be0b05afed05f37e0e3c8c03969ee1a94a0d