Add MIME_TYPE filters in Search APIs documentation
Add MIME_TYPES filter in the documentation for cloud picker's search
APIs. This will indicate to the CMP that a certain mime type filter has
been applied to a PhotoPicker session and the results should not contain
media items of the given mime type.

Also update the documentation of onQueryAlbums which is an existing API
that we send MIME type filters to, however the documentation did not
reflect this.

Bug: 374927525
Test: NA
Flag: com.android.providers.media.flags.cloud_media_provider_search
Change-Id: I949e546bf2d3835a8f1ecd29307228caccf10e7c